requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
colorama==0.4.1
coverage==4.5.3
entrypoints==0.3
flake8==3.7.7
mccabe==0.6.1
more-itertools==6.0.0
pluggy==0.9.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
six==1.12.0
allure-pytest
allure-python-commons
pytest
pytest-xdist
selenium
webdriver_manager